requirements.txt
aiohttp>=3.7.2
discord>=1.0.1
validators>=0.18.1
setuptools>=50.3.2
captcha>=0.3
emoji>=0.6.0
tortoise-orm==0.16.18